Update Provision Response and fix float temperature value

+++ src/main/java/com/dell/dtc/edge/guard/model/TemperatureData.java
@@ -16,5 +16,5 @@ public class TemperatureData {
     private String orgId;
 
     @JsonProperty("temperature")
-    private float temperature;
+    private int temperature;
 }
